% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conan.R
\docType{package}
\name{conan-package}
\alias{conan}
\alias{conan-package}
\title{For us, there is no spring. Just the wind that smells fresh before
the storm.}
\description{
Conan the Librarian
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrc-ide/conan}
  \item Report bugs at \url{https://github.com/mrc-ide/conan/issues}
}

}
\author{
\strong{Maintainer}: Rich FitzJohn \email{rich.fitzjohn@gmail.com}

Other contributors:
\itemize{
  \item Imperial College of Science, Technology and Medicine [copyright holder]
}

}
