% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conan.R
\name{conan}
\alias{conan}
\title{Conan the Librarian}
\usage{
conan(filename, packages, repos = NULL, policy = "upgrade")
}
\arguments{
\item{filename}{Path to write the install script to. Any directory
components will be created as needed.}

\item{packages}{A character vector of packages to install. These
can be names of cran packages or github references etc; see
\code{\link[pkgdepends:pkg_installation_proposal]{pkgdepends::new_pkg_installation_proposal()}} for more details}

\item{repos}{A character vector of repositories to use when
installing. A suitable CRAN repo will be added if not detected.}

\item{policy}{Should be either "lazy" or "upgrade", with a default
of "upgrade"}
}
\value{
Invisibly, the path to the created script. This can be run
via \code{Rscript} to create a standalone library. The script will
require exactly one argument: the path to the library to
create. It will respond to the environment variables
\code{CONAN_PATH_BOOTSTRAP} and \code{CONAN_PATH_CACHE}.
}
\description{
High-level conan interface. This function creates a standalone
script that can provision a library at any location. The script
has no dependencies and can be copied to another system and create
a library given an empty R installation.
}
\examples{
path <- conan::conan(tempfile(), "cpp11")
writeLines(tail(readLines(path)))
}
