% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{conan_bootstrap}
\alias{conan_bootstrap}
\title{Create a bootstrap library}
\usage{
conan_bootstrap(path, upgrade = FALSE)
}
\arguments{
\item{path}{Path to create library}

\item{upgrade}{Logical, indicating if the bootstrap packages
should be upgraded if newer versions are available.}
}
\value{
Invisibly, \code{path}, but primarily called for its side
effects, which is to create or update the library at \code{path}
}
\description{
Create a bootstrap library
}
\examples{
if (interactive()) {
  path <- tempfile()
  conan::conan_bootstrap(path)
  dir(path)
}
}
