% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{conan_check}
\alias{conan_check}
\title{Check if packages in library}
\usage{
conan_check(packages, library)
}
\arguments{
\item{packages}{A character vector of package names or pkgdepends
"references"}

\item{library}{A path to a library}
}
\value{
A list with elements
\itemize{
\item \code{complete}: logical, indicating if all packages were found
\item \code{found}: A character vector of found packages
\item \code{missing}: A character vector of missing packages
}
}
\description{
Check to see if all packages are available in a given library
}
\examples{
# Simple usage:
conan::conan_check(c("conan", "pkgdepends"), .libPaths())

# While we parse references, we don't check version information:
conan::conan_check("github::mrc-ide/conan@v2.0.0", .libPaths())

# Missing packages will be returned as the inferred package name
conan::conan_check("github::org/unknownpkg", .libPaths())
}
