% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryrun.R
\name{conan_dryrun}
\alias{conan_dryrun}
\title{Dry run of package installation}
\usage{
conan_dryrun(
  packages,
  policy = "upgrade",
  repos = NULL,
  cran = NULL,
  lib = NULL,
  error = TRUE
)
}
\arguments{
\item{packages}{A character vector of packages to install. These
can be names of CRAN packages or GitHub references etc; see
\code{\link[pkgdepends:pkg_installation_proposal]{pkgdepends::new_pkg_installation_proposal()}} for more details}

\item{policy}{Should be either "lazy" or "upgrade", with a default
of "upgrade"}

\item{repos}{A character vector of repositories to use when
installing. A suitable CRAN repo will be added if not detected
(using the \code{cran} argument if provided)}

\item{cran}{Fallback CRAN repo to use. If not given we will use
\verb{https://cloud.r-project.org}}

\item{lib}{The path to install into. We will create a
self-contained library at this address.}

\item{error}{Logical, indicating if error should be thrown on
failure. If \code{FALSE} then we return the proposal object instead.}
}
\value{
Invisibly, the resolution
}
\description{
Dry run of package installation.
}
\examples{
\dontshow{if (nzchar(Sys.getenv("R_USER_CACHE_DIR"))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
conan::conan_dryrun("cpp11")
\dontshow{\}) # examplesIf}
}
