% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{conan_scripts}
\alias{conan_scripts}
\title{Write scripts to directory}
\usage{
conan_scripts(path)
}
\arguments{
\item{path}{Path to write to}
}
\value{
Nothing, called for its side effects
}
\description{
Write conan's scripts to a directory
}
\examples{
path <- tempfile()
conan::conan_scripts(path)
dir(path)
}
