% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{conan_sources}
\alias{conan_sources}
\title{Describe sources}
\usage{
conan_sources(packages, repos = NULL, cran = NULL)
}
\arguments{
\item{packages}{A character vector of packages to install. These
can be names of cran packages or github references etc; see
\code{\link[pkgdepends:pkg_installation_proposal]{pkgdepends::new_pkg_installation_proposal()}} for more details}

\item{repos}{A character vector of repositories to use when
installing. A suitable CRAN repo will be added if not detected
(using the \code{cran} argument if provided)}

\item{cran}{Fallback CRAN repo to use. If not given we will use
\verb{https://cloud.r-project.org}}
}
\value{
An object of type \code{conan_sources} which can be used later
with \link{conan}, \link{conan_install} etc
}
\description{
Describe sources
}
\examples{
conan::conan_sources(c("pkg1", "pkg2"))
}
\author{
Rich FitzJohn
}
