% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{conan_write}
\alias{conan_write}
\title{Write conan installation script}
\usage{
conan_write(config, path)
}
\arguments{
\item{config}{Conan config, from \code{\link[=conan_configure]{conan_configure()}}}

\item{path}{The path to write to}
}
\value{
Nothing
}
\description{
Write a conan installation script
}
