% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_t_exprs.R
\name{cor_t_exprs}
\alias{cor_t_exprs}
\title{Expression Correlation}
\usage{
cor_t_exprs(eset)
}
\arguments{
\item{eset}{An expression set object}
}
\value{
A single numeric value
}
\description{
Calculates the correlation of genes across samples from an expression set object.
}
\keyword{internal}
