% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_conn.R
\name{diff_conn}
\alias{diff_conn}
\title{Differential Connectivity}
\usage{
diff_conn(eset, mod_list, covariate, ctrl, cond,
  sim_type = c("bootstrap", "permutation"), iter = 5,
  mean_correct = FALSE, cores = 1, use_gpu = FALSE,
  mdc_type = c("frac", "diff"), plotting = FALSE, reporting = FALSE,
  report_dir = "./data")
}
\arguments{
\item{eset}{An expression set object}

\item{mod_list}{A list of character vectors}

\item{covariate}{A string corresponding to column in pData distinguishing sample groups}

\item{ctrl}{A string defining control samples}

\item{cond}{A string defining condition samples}

\item{sim_type}{Simulation type can be either c("bootstrap", "permutation")}

\item{iter}{Number of sampling iterations during significance testing}

\item{mean_correct}{Correct with background connectivity}

\item{cores}{Number of CPU cores available for parallelization}

\item{use_gpu}{Use of GPU computing}

\item{mdc_type}{Method for calculating difference in connectivity can be either c("frac", "diff")}

\item{reporting}{Generate a markdown report for analysis}

\item{report_dir}{Directory where report is generated}
}
\value{
A list of statistics and plots resulting from the analysis
}
\description{
Calculates the differential connectivity between two groups of samples across one or more modules of genes.
}
