% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_iter.R
\name{do_iter}
\alias{do_iter}
\title{Iteration of Random Sampling}
\usage{
do_iter(n, mod_list, r_eset, t_eset, sim_type, mean_correct, mdc_type)
}
\arguments{
\item{n}{An unused placeholder variable}

\item{mod_list}{A list of modules of gene names}

\item{r_eset}{Expression set of control samples}

\item{t_eset}{Expression set of condition samples}

\item{sim_type}{Simulation type can be either c("bootstrap", "permutation")}

\item{mean_correct}{Correct with background connectivity}

\item{mdc_type}{Method for calculating difference in connectivity can be either c("frac", "diff")}
}
\value{
Connectivity statistics for the iteration
}
\description{
Performs an iteration random sampling and calculating differential connectivity
}
\keyword{internal}
