% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_plot.R
\name{do_plot}
\alias{do_plot}
\title{Plot Module Differential Connectivity}
\usage{
do_plot(output, mean_correct)
}
\arguments{
\item{output}{Output of diff_conn}

\item{mean_correct}{Logical. Was mean correct used in diff_conn?}
}
\value{
A list of ggplots
}
\description{
Plot Module Differential Connectivity
}
