% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_novar.R
\name{genes_novar}
\alias{genes_novar}
\title{No-Variance Genes}
\usage{
genes_novar(r_eset, t_eset)
}
\arguments{
\item{r_eset}{Expression set of control samples}

\item{t_eset}{Expression set of condition samples}
}
\value{
A vector of gene names with no variance
}
\description{
Finds genes without variance across either control or condition samples
}
\keyword{internal}
