% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cvs.R
\name{get_cvs}
\alias{get_cvs}
\title{Get Connectivity Vectors}
\usage{
get_cvs(mod_genes, r_eset, t_eset)
}
\arguments{
\item{mod_genes}{A character vector of genes within a module}

\item{r_eset}{Expression set for the reference samples}

\item{t_eset}{Expression set for the test samples}
}
\value{
A list of connectivity vectors
}
\description{
Get Connectivity Vectors
}
\keyword{internal}
