% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_upper_tri}
\alias{get_upper_tri}
\title{Get Upper Triangle}
\usage{
get_upper_tri(mat, diag)
}
\arguments{
\item{mat}{A matrix.}

\item{diag}{Logical. Should the diagonal be included?}
}
\value{
A vector for values.
}
\description{
Returns the upper triangle of a matrix in the form of a vector.
}
\keyword{internal}
