% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lower_tri}
\alias{lower_tri}
\title{Get Lower Triangle}
\usage{
lower_tri(mat, diag)
}
\arguments{
\item{mat}{A matrix.}

\item{diag}{Logical. Should the diagonal be included?}
}
\value{
A vector for values.
}
\description{
Returns the lower triangle of a matrix in the form of a vector.
}
\keyword{internal}
