% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mdc.R
\name{plot_mdc}
\alias{plot_mdc}
\title{Plot Module Differential Connectivity}
\usage{
plot_mdc(output, mean_correct)
}
\arguments{
\item{output}{Output of diff_conn}

\item{mean_correct}{Logical. Was mean correct used in diff_conn?}
}
\value{
A list of ggplots
}
\description{
Plot Module Differential Connectivity
}
