% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConAn.r
\docType{package}
\name{ConAn-package}
\alias{ConAn}
\alias{ConAn-package}
\title{ConAn: Differential Network Connectivity Analysis}
\description{
A package to measure differential connectivity of multiple networks between two conditions with visualization and robust statistical testing.
}
