% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conan.R
\name{conan}
\alias{conan}
\title{Differential Connectivity}
\usage{
conan(
  eset,
  mod_list,
  covariate,
  ctrl,
  cond,
  sim_type = c("bootstrap", "permutation"),
  iter = 5,
  mean_correct = FALSE,
  N_genes = NULL,
  iter_bg = 1,
  cores = 1,
  mdc_type = c("difference", "fraction"),
  plotting = FALSE,
  FUN = cor,
  ...
)
}
\arguments{
\item{eset}{An expression set object}

\item{mod_list}{A list of character vectors}

\item{covariate}{A string corresponding to column in pData distinguishing sample groups}

\item{ctrl}{A string defining control samples}

\item{cond}{A string defining condition samples}

\item{sim_type}{Simulation type can be either c("bootstrap", "permutation")}

\item{iter}{Number of sampling iterations during significance testing}

\item{mean_correct}{Correct with background connectivity}

\item{N_genes}{Number of randomly selected genes to be used during each iteration}

\item{iter_bg}{Number of iterations used when calculating background (only used for non-NULL N_genes values}

\item{cores}{Number of cores available for parallelization}

\item{mdc_type}{Method for calculating difference in connectivity can be either c("fraction", "difference")}

\item{FUN}{Function for calculating correlation}
}
\value{
A list of statistics and plots resulting from the analysis
}
\description{
Calculates the differential connectivity between two groups of samples across one or more modules of genes.
}
