% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch.R
\name{conan_watch}
\alias{conan_watch}
\title{Watch a conan installation}
\usage{
conan_watch(
  get_status,
  get_log,
  show_progress = TRUE,
  show_log = TRUE,
  poll = 1,
  error = TRUE
)
}
\arguments{
\item{get_status}{A callback to get the status of the
installation. The callback will be called with no arguments and
must return \code{COMPLETE} on successful completion, \code{ERROR} on
failure and any other value to continue.}

\item{get_log}{A callback to read logs of the installation
(something like \code{function() readLines(filename, warn = FALSE)}
may be sufficient)}

\item{show_progress}{Logical, indicating if we should show a
progress bar. This requires the progress package.}

\item{show_log}{Logical, indicating if the installation log should
be printed}

\item{poll}{Time, in seconds, used to throttle calls to the status
function. The default is 1 second}

\item{error}{Logical, indicating if we should throw an error if
installation fails.}
}
\value{
The final status (\code{COMPLETE} or \code{ERROR})
}
\description{
Watch a conan installation from another process. This can be used
to stream the status and progress back to the current R process.
}
\author{
Richard Fitzjohn
}
