% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.R
\name{do_sampling}
\alias{do_sampling}
\title{Randomly Shuffle Samples}
\usage{
do_sampling(
  iter,
  c_samples,
  r_samples,
  t_samples,
  method = c("bootstrap", "permutation")
)
}
\arguments{
\item{c_samples}{Combined control and condition sample names}

\item{r_samples}{Control sample names}

\item{t_samples}{Condition sample names}

\item{method}{Sampling method can be either c("bootstrap", "permutation")}
}
\value{
Return shuffled samples for reference and test expression sets
}
\description{
Randomly Shuffle Samples
}
\keyword{internal}
